server/config/config.py
import os

from dotenv import load_dotenv

load_dotenv()

class Config:
    SECRET_KEY = os.getenv("SECRET_KEY")
    GOOGLE_SECRET = os.getenv("GOOGLE_SECRET")
    GOOGLE_CLIENT_ID = os.getenv("GOOGLE_CLIENT_ID")
    JSON_SORT_KEYS = False # Speed up jsonify
